# api/openapi.yaml
openapi: 3.0.3
info:
  title: sshc Local API
  version: 0.1.0
paths:
...
  /api/v1/terminal/backgrounds:
    get:
      operationId: listTerminalBackgrounds
      responses:
        "200":
          description: The images this workspace holds
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TerminalBackgroundList" }
        "401": { $ref: "#/components/responses/Problem" }
    post:
      operationId: addTerminalBackground
      # クライアントは希望名と内容を送り、サーバーが保存名と型を決定する。
      parameters:
        - in: query
          name: name
          required: false
          schema: { type: string, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        "201":
          description: The image was stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TerminalBackground" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
  /api/v1/terminal/backgrounds/capacity:
    put:
      operationId: setTerminalBackgroundCapacity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TerminalBackgroundCapacityRequest" }
      responses:
        "200":
          description: The updated image library and capacity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TerminalBackgroundList" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/terminal/backgrounds/{name}:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string, maxLength: 128 }
    get:
      operationId: getTerminalBackground
      responses:
        "200":
          description: The image itself
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      operationId: renameTerminalBackground
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RenameTerminalBackgroundRequest" }
      responses:
        "200":
          description: The image and its saved references were renamed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TerminalBackground" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteTerminalBackground
      responses:
        "204": { description: The image is gone }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        detail: { type: string }
        path: { type: string }
        line: { type: integer }
        column: { type: integer }
        diagnostics: { type: array, items: { $ref: "#/components/schemas/Diagnostic" } }
        conflict: { $ref: "#/components/schemas/ConflictReport" }
        blockers: { type: array, items: { type: string } }
        currentVersion: { type: integer, minimum: 0 }
        requiredVersion: { type: integer, minimum: 1 }
        # blockers は拒否理由を安定したコードで返す。
        # field・reason・limit は、入力のどの項目をどの理由で受け取れなかったかを
        # 決まった語で返す。field は保存形式の JSON のパス、limit は長さや数の上限。
        # VPN の経路や接続先へ届かなかった応答（vpn_session_failed、vpn_target_failed、
        # vpn_destination_invalid）では reason だけが入る。
        field: { type: string }
        reason: { type: string }
        limit: { type: integer, minimum: 1 }
        # 設定ファイル（OpenVPN の .ovpn）の中の誤りでは、断った指示の名前が入る。行番号は
        # line に入る。どちらも engine が決めた語と数だけで、設定ファイルの中身は載せない。
        directive: { type: string }
...
    Diagnostic:
      type: object
      additionalProperties: false
      required: [severity, code]
      properties:
        severity: { type: string }
        code: { type: string }
        path: { type: string }
        absolute: { type: string }
        external: { type: boolean }
        line: { type: integer }
        detail: { type: string }
...
    TerminalBackground:
      type: object
      additionalProperties: false
      required: [name, bytes, type]
      properties:
        name: { type: string }
        bytes: { type: integer }
        type: { type: string }
    RenameTerminalBackgroundRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
    TerminalBackgroundCapacityRequest:
      type: object
      additionalProperties: false
      required: [capacityMiB]
      properties:
        capacityMiB: { type: integer, minimum: 1, maximum: 1024 }
    TerminalBackgroundList:
      type: object
      additionalProperties: false
      required: [backgrounds, usedBytes, capacityBytes, remainingBytes]
      properties:
        backgrounds: { type: array, items: { $ref: "#/components/schemas/TerminalBackground" } }
        usedBytes: { type: integer }
        capacityBytes: { type: integer }
        # 追加保存できる残りバイト数。
        remainingBytes: { type: integer }
...
    DiffLine:
      type: object
      additionalProperties: false
      required: [op, text]
      properties:
        op: { type: string }
        text: { type: string }
        oldLine: { type: integer }
        newLine: { type: integer }
...
    ConflictReport:
      type: object
      additionalProperties: false
      required: [path, externalChange, localChange]
      properties:
        path: { type: string }
        baseDigest: { type: string }
        diskDigest: { type: string }
        externalChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
        localChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
...
  responses:
    Problem:
      description: Request rejected
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }